+++ spec/spec_helper.rb
@@ -3,9 +3,12 @@ SimpleCov.start do
   add_filter 'spec/'
 end
 
+require 'timecop'
 require 'pry'
 require_relative 'test_object'
 require_relative '../lib/cache_store_redis'
+
+Timecop.safe_mode = true
 
 RSpec.configure do |config|
   config.expect_with :rspec do |expectations|