+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Dropped Support for JRuby
 * Improved handling of expiry when calling `set`.
+* Connection Pooling (Added in v1.1.0).